Transform a ggkegg graph to igraph or visNetwork

Transform a ggkegg graph to igraph or visNetwork

Usage

kegg_to_graph(path_id, return_type = "igraph", scaling_factor = 1.5)

Arguments

path_id

KEGG pathway ID (e.g., "hsa:04110" or "04110").

return_type

Output type: "igraph" or "visNetwork".

scaling_factor

Numeric factor to scale node sizes.

Value

An igraph or visNetwork object representing the pathway.

Details

This function downloads the KGML file for the specified KEGG pathway, then parses it to generate a graph representation using either the igraph or visNetwork package. It styles nodes and edges based on their types the output can be used for visualization or further analysis. If differential expression results are provided, they can be mapped to the nodes using the function map_results_to_graph.

Examples

pathway <- "hsa04110"  # Example pathway ID
graph <- kegg_to_graph(pathway)
#> Downloading KGML for hsa04110 ...
#> Downloaded & cached: hsa04110
#> Parsed 134 nodes from KGML file.
#> Parsed 119 edges from KGML file.
kegg_to_graph(pathway, return_type = "visNetwork")
#> Using cached KEGG KGML for hsa04110
#> Parsed 134 nodes from KGML file.
#> Parsed 119 edges from KGML file.